Fix percentage (#221)
* alumni status bar should now give correct percentage

* now it should work

+++ src/DevBetterWeb.Core/Services/MemberSubscriptionPeriodCalculationsService.cs
@@ -6,7 +6,7 @@ namespace DevBetterWeb.Core.Services
 {
   public class MemberSubscriptionPeriodCalculationsService : IMemberSubscriptionPeriodCalculationsService
   {
-    private const int DAYS_SUBSCRIBED_TO_BECOME_ALUMNUS = 730;
+    private const double DAYS_SUBSCRIBED_TO_BECOME_ALUMNUS = 730;
 
     public bool GetHasCurrentSubscription(Member member)
     {
